% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TephraMap.R
\name{tephra.estimate}
\alias{tephra.estimate}
\title{Estimate the isopachs and total volume given tephra
samples.}
\usage{
tephra.estimate(
  fn,
  coordsys,
  UTMtz = NaN,
  super = FALSE,
  prior.range = NULL,
  prior.sigma = NULL
)
}
\arguments{
\item{fn}{String of the csv file containing the sampled tephra deposits. This must have columns containing the X coordinate, Y coordinate and thickness of the tephra deposit. There must be headers for all columns, with the sampled thickness having the units as the header: <mm>, <cm> or <m> (without the < >).}

\item{coordsys}{String of the coordinate system used. This must be either \code{"UTM"}, which is the UTM coordinates (where the UTM zone must also be given); \code{"dfs"}, which is distance from some origin in kilometres; or \code{"latlon"} for WG84 latitude and longitude, where the UTM zone must also be supplied as units are converted to UTM if \code{super=FALSE}.}

\item{UTMtz}{(optional) The UTM zone if using UTM or latlon coordinates (and \code{super==FALSE})}

\item{super}{(optional, default=\code{FALSE}) Set to true if this is a super eruption. The calculation is then done on a globe rather than a flat plain. Coordinates must be in latlon. (This only works for super eruptions, i.e. those spanning several degrees longitude/latitude).}

\item{prior.range}{(optional, default=\code{NULL}) Prior probability for the spatial range of the tephra field in km using a penalised complexity prior. This is a vector of length 2, \code{c(range0, Prange)}, where Pr(range < range0) = Prange. Setting Prange as \code{NA} fixes range to range0. Leave as \code{NULL} for the default prior, approximated from the measurement field.}

\item{prior.sigma}{(optional, default=\code{NULL}) Prior probability for the standard deviation of the log-tickness of the tephra field defined using a penalised complexity prior. This is a vector of length 2, \code{c(sigma0, Psigma)}, where Pr(sigma > sigma0) = Psigma. Setting Psigma as \code{NA} fixes sigma to sigma0. Leave as \code{NULL} for default prior.}
}
\value{
A list containing (1) \code{$out}: a list as S3 object. The list contains the data.frame \code{$data} with the x and y coordinates and value of input data in mm; \code{$field}, a list containing the estimated tephra thickness field on a 500x500 grid in mm and the associated coordinates; and \code{$volume} the estimated total volume in km\eqn{^3}. (2) \code{$plotparams}, a list containing some parameters for plotting
}
\description{
Estimate the isopachs and total volume given tephra
samples.
}
