% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TephraMap.R
\name{tephra.map}
\alias{tephra.map}
\title{Estimate isopachs and total volume}
\usage{
tephra.map(
  datafile,
  coord,
  UTMzone = NaN,
  plot = TRUE,
  super = FALSE,
  prior.range = NULL,
  prior.sigma = NULL
)
}
\arguments{
\item{datafile}{String of the csv file containing the sampled tephra deposits. This must have columns containing the X coordinate, Y coordinate and thickness of the tephra deposit. There must be headers for all columns, with the sampled thickness having the units as the header: <mm>, <cm> or <m> (without the < >).}

\item{coord}{String of the coordinate system used. This must be either \code{"UTM"}, which is the UTM coordinates (where the UTM zone must also be given); \code{"dfs"}, which is distance from some origin in kilometres; or \code{"latlon"} for WG84 latitude and longitude, where the UTM zone must also be supplied as units are converted to UTM if \code{super=FALSE}.}

\item{UTMzone}{(optional) The UTM zone if using UTM or latlon coordinates (and \code{super==FALSE})}

\item{plot}{(optional, default=\code{TRUE}) whether to plot the output}

\item{super}{(optional, default=\code{FALSE}) Set to \code{TRUE} if this is a super eruption. The calculation is then done on a globe rather than a flat plain. Coordinates must be in latlon. (This only works for super eruptions, i.e. those spanning several degrees longitude/latitude)}

\item{prior.range}{(optional, default=\code{NULL}) Prior probability for the spatial range of the tephra field in km using a penalised complexity prior. This is a vector of length 2, \code{c(range0, Prange)}, where Pr(range < range0) = Prange. Setting Prange as \code{NA} fixes range to range0. Leave as \code{NULL} for the default prior, approximated from the measurement field.}

\item{prior.sigma}{(optional, default=\code{NULL}) Prior probability for the standard deviation of the log-tickness of the tephra field defined using a penalised complexity prior. This is a vector of length 2, \code{c(sigma0, Psigma)}, where Pr(sigma > sigma0) = Psigma. Setting Psigma as \code{NA} fixes sigma to sigma0. Leave as \code{NULL} for default prior.}
}
\value{
List as S3 object. List contains: \code{$data}: a data.frame  with the x and y coordinates and value of input data in mm; \code{$field}: a list containing the estimated tephra thickness on a 500x500 grid in mm and the associated coordinates; and \code{$volume}: the estimated total volume in km\eqn{^3}.
}
\description{
Estimate the isopachs and total volume given tephra
samples and optionally plot the result. The estimated
tephra field is returned on a 500x500 rasterised grid
with the corresponding coordinates and an estimate of
the total volume and the input data.
}
\examples{
 tephrafield <- tephra.map("CerroNegro.csv", "UTM", UTMzone=16)
}
