% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TephraMap.R
\name{tephra.plot}
\alias{tephra.plot}
\title{Plot an isopach map}
\usage{
tephra.plot(outlist)
}
\arguments{
\item{outlist}{The output from \code{tephra.estimate()}}
}
\description{
Plot an isopach map after estimating the thickness
}
